✨ (env): add COMPOSE_BAKE variable to .env.example 🔧 (husky): comment out commitlint in commit-msg hook 🔄 (install_1password): specify version for 1Password download script
Add the `COMPOSE_BAKE` variable to the `.env.example` file to provide a new configuration option. Comment out the commitlint command in the Husky commit-msg hook to prevent it from running. Update the 1Password installation script to specify the version being downloaded, ensuring compatibility and clarity in the installation process.